Move helper functions to own header file
